requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.9.1
blinker==1.9.0
cachetools==5.5.1
certifi==2024.7.4
charset-normalizer==3.4.1
click==8.1.8
configparser==7.1.0
contourpy==1.3.2
crispy-bootstrap4==2025.6
crispy-bootstrap5==2025.6
cycler==0.12.1
distlib==0.4.0
distro==1.9.0
Django==5.2.6
django-crispy-forms==2.4
django-tastypie==0.15.1
django-widget-tweaks==1.5.0
filelock==3.19.1
Flask==3.1.0
fonttools==4.57.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.67.0
grpcio==1.76.0rc1
grpcio-status==1.71.2
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.5
kiwisolver==1.4.8
macholib==1.16.3
MarkupSafe==3.0.2
matplotlib==3.10.1
modulegraph==0.19.6
mpmath==1.3.0
numpy==2.2.5
oauthlib==3.2.2
openai==1.35.15
packaging==25.0
pillow==11.2.1
pipenv==2025.0.4
platformdirs==4.4.0
priority==2.0.0
proto-plus==1.26.0
protobuf==5.29.3
py2app==0.28.8
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.8.2
pydantic_core==2.20.1
pygame==2.6.1
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.3
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-mimeparse==2.0.0
requests==2.32.5
rsa==4.9
segno==1.6.1
setuptools==80.0.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
sympy==1.14.0
tqdm==4.66.4
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
virtualenv==20.34.0
Werkzeug==3.1.3
wheel==0.45.1
wsproto==1.2.0
